This closes #4, closes #5
Fix the issue 4 by adding the taglib library.
Fix the issue 5 by checking if the file exists in the temp directory.

+++ Musics - Client/Player.cs
@@ -27,9 +27,12 @@ namespace Musics___Client
 
             string path = @"c:\MusicsFiles\" + music.Title + music.Format;
 
-            
 
-            File.WriteAllBytes(path, music.FileBinary);
+            if (!File.Exists(path))
+            {
+                File.WriteAllBytes(path, music.FileBinary);
+            }
+            
 
             player.URL = path;
             player.controls.play();
